Implementação do Spring Security, criação e gerenciamento de perfis, Tokens JWT

+++ src/main/java/com/biblioteca_icpi/service/LivroService.java
@@ -1,9 +1,12 @@
 package com.biblioteca_icpi.service;
 
-import com.biblioteca_icpi.dto.LivroDTO;
+import com.biblioteca_icpi.dto.CadastrarLivroDTO;
+import com.biblioteca_icpi.dto.LivroResponseDTO;
 import com.biblioteca_icpi.exception.livro.LivroJaExistenteException;
 import com.biblioteca_icpi.exception.livro.LivroNaoEncontradoException;
+import com.biblioteca_icpi.model.DisponibilidadeLivro;
 import com.biblioteca_icpi.model.Livro;
+import com.biblioteca_icpi.repository.AluguelRepository;
 import com.biblioteca_icpi.repository.LivroRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -14,18 +17,20 @@ import java.util.Optional;
 public class LivroService {
 
     private final LivroRepository livroRepository;
+    private  final AluguelRepository aluguelRepository;
 
-    public LivroService(LivroRepository livroRepository) {
+    public LivroService(LivroRepository livroRepository, AluguelRepository aluguelRepository) {
         this.livroRepository = livroRepository;
+        this.aluguelRepository = aluguelRepository;
     }
 
-    public LivroDTO buscarLivroNoBancoDeDados (Long idLivro) {
+    public LivroResponseDTO buscarLivroNoBancoDeDados (Long idLivro) {
         Livro livro = livroRepository.findById(idLivro).orElseThrow(() -> new LivroNaoEncontradoException("Livro não encontrado!"));
-        return convertToLivroDTO(livro);
+        return convertToLivroResponseDTO(livro);
     }
 
     @Transactional
-    public LivroDTO cadastrarLivro (LivroDTO dto) {
+    public LivroResponseDTO cadastrarLivro (CadastrarLivroDTO dto) {
         Optional<Livro> possivelLivro = livroRepository.findByNome(dto.getNome());
         if (possivelLivro.isPresent()) {
             throw new LivroJaExistenteException("Livro já cadastrado!");
@@ -36,36 +41,40 @@ public class LivroService {
             livro.setAutor(dto.getAutor());
             livro.setDescricao(dto.getDescricao());
             livro.setGenero(dto.getGenero());
+            livro.setDisponivel(DisponibilidadeLivro.DISPONIVEL);
             livroRepository.save(livro);
-            return convertToLivroDTO(livro);
+            return convertToLivroResponseDTO(livro);
         }
     }
 
     @Transactional
-    public LivroDTO editarLivro (Long idLivro, LivroDTO dto) {
+    public LivroResponseDTO editarLivro (Long idLivro, CadastrarLivroDTO dto) {
         Livro livroEncontrado = livroRepository.findById(idLivro).orElseThrow(() -> new LivroNaoEncontradoException("Livro não encontrado!"));
         livroEncontrado.setAutor(dto.getAutor());
         livroEncontrado.setNome(dto.getNome());
         livroEncontrado.setDescricao(dto.getDescricao());
         livroEncontrado.setGenero(dto.getGenero());
         livroRepository.save(livroEncontrado);
-        return convertToLivroDTO(livroEncontrado);
+        return convertToLivroResponseDTO(livroEncontrado);
     }
 
     public void excluirLivro (Long idLivro) {
         Livro livroEncontrado = livroRepository.findById(idLivro).orElseThrow(() -> new LivroNaoEncontradoException("Livro não encontrado!"));
+        if (aluguelRepository.existsByLivro(livroEncontrado)) {
+            throw new IllegalStateException("O livro não pode ser deletado pois possui histórico de alugueis");
+        }
         livroRepository.delete(livroEncontrado);
     }
 
-    public List<LivroDTO> listarLivros () {
+    public List<LivroResponseDTO> listarLivros () {
         List<Livro> livros = livroRepository.findAll();
-        return livros.stream().map(this::convertToLivroDTO).toList();
+        return livros.stream().map(this::convertToLivroResponseDTO).toList();
     }
 
 
     public void marcarComoAlugado (Livro livro) {
-        if (livro.isDisponivel()) {
-            livro.setDisponivel(false);
+        if (livro.getDisponivel() == DisponibilidadeLivro.DISPONIVEL) {
+            livro.setDisponivel(DisponibilidadeLivro.ALUGADO);
             livroRepository.save(livro);
         }
         else {
@@ -74,21 +83,23 @@ public class LivroService {
     }
 
     public void marcarComoDisponivel (Livro livro) {
-        if (livro.isDisponivel()) {
+        if (livro.getDisponivel() == DisponibilidadeLivro.DISPONIVEL) {
             throw  new IllegalStateException("Este livro já está disponível!");
         }
         else {
-            livro.setDisponivel(true);
+            livro.setDisponivel(DisponibilidadeLivro.DISPONIVEL);
             livroRepository.save(livro);
         }
     }
 
-    public LivroDTO convertToLivroDTO (Livro livro) {
-        LivroDTO livroDTO = new LivroDTO();
-        livroDTO.setNome(livro.getNome());
-        livroDTO.setAutor(livro.getAutor());
-        livroDTO.setDescricao(livro.getDescricao());
-        livroDTO.setGenero(livro.getGenero());
-        return livroDTO;
+    public LivroResponseDTO convertToLivroResponseDTO (Livro livro) {
+        LivroResponseDTO livroResponseDTO = new LivroResponseDTO();
+        livroResponseDTO.setId(livro.getId());
+        livroResponseDTO.setNome(livro.getNome());
+        livroResponseDTO.setAutor(livro.getAutor());
+        livroResponseDTO.setDescricao(livro.getDescricao());
+        livroResponseDTO.setGenero(livro.getGenero());
+        livroResponseDTO.setStatus(livro.getDisponivel());
+        return livroResponseDTO;
     }
 }
